+++ CHANGES.md
@@ -3,6 +3,7 @@
 - update LaTeX backends 
 - LaTeX tabular mode now uses `tabularray`
 - add a new LaTeX mode using `simplebnf`
+- remove `suffix` dependency in favor of `xparse`
 
 ## v0.7.0 - 2024-07-03
 - support for EBNF output ([#14](https://github.com/Lelio-Brun/Obelisk/issues/14))